% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-recodes-and-helpers.R
\name{acs.commute.recode}
\alias{acs.commute.recode}
\title{q.recode.commute.table.B08006}
\usage{
acs.commute.recode(x, separate.carpools = F, filter.aggregates = T)
}
\arguments{
\item{separate.carpools}{whether to keep carpoolers separate}

\item{filter.aggregates}{If true, filter totals and subtotals.}

\item{commutes}{a pull of table B08006, as from \code{multiyr.acs.wrapper(...)}.}
}
\description{
Uses table B08006, SEX OF WORKERS BY MEANS OF TRANSPORTATION TO WORK.
Immediately trims the breakdown by sex.
}
