% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-commutes.R
\name{acs.commute.share.recode}
\alias{acs.commute.share.recode}
\title{acs.commute.share.recode}
\usage{
acs.commute.share.recode(
  commutes,
  separate.carpools = F,
  filter.aggregates = T
)
}
\arguments{
\item{commutes}{a pull of table B08006, as from \code{multiyr.acs.wrapper(...)}.}

\item{separate.carpools}{whether to keep carpoolers separate}

\item{filter.aggregates}{If true, filter totals and subtotals.}
}
\description{
Uses table B08006, SEX OF WORKERS BY MEANS OF TRANSPORTATION TO WORK.
Immediately trims the breakdown by sex.
}
