% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-recodes-and-helpers.R
\name{acs.rentals.by.level.recode}
\alias{acs.rentals.by.level.recode}
\title{acs.rentals.by.level.recode}
\usage{
acs.rentals.by.level.recode(
  x,
  acs.tbl = c("B25056", "B25061"),
  filter.aggregates = T
)
}
\description{
Recodes table B25056 (universe: renter-occupied housing units) or table
B25061 (universe: Vacant-for-rent and Rented, Not Occupied Housing Units)
}
