% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-recodes-and-helpers.R
\name{acs.tenure.by.hh.size.recode}
\alias{acs.tenure.by.hh.size.recode}
\title{acs.tenure.by.hh.size.recode}
\usage{
acs.tenure.by.hh.size.recode(x)
}
\description{
Recode table B25009 (unv is occupied housing units); separates out tenure
from size.
}
