% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acs-commutes.R
\name{commute.times.by.mode}
\alias{commute.times.by.mode}
\title{commute.times.by.mode}
\usage{
commute.times.by.mode(geo = "county", state, cofps = NULL, year = 2019)
}
\arguments{
\item{geo}{geography type, passed to tidycensus. Default is county. Not
not all geographies available for this pull.}

\item{state}{state fp or abrv.}

\item{year}{year}
}
\description{
Rigid function: outputs a table for counties or other (larger) areas in a
state with modeshare and avg. commute times for bus/car/trains/and total
transit.
}
