% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-helpers.R
\name{extract.acs.var}
\alias{extract.acs.var}
\title{extract.acs.var}
\usage{
extract.acs.var(x)
}
\description{
From acs table with Btable#_var# format, extract the variable. Takes the
column called "variable" for table pulled with tidycensus.
}
