% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-mass-pulls.R
\name{get.all.rentals.by.price.lvl}
\alias{get.all.rentals.by.price.lvl}
\title{get.all.rentals.by.price.lvl}
\usage{
get.all.rentals.by.price.lvl(states, years, geo, survey = "acs5")
}
\description{
ACS by default includes "vacant" and "occupied" housing units separately in
many tables. This function wraps the recode functions for rentals by price
tables:
}
\details{
\itemize{
\item B25056 (universe: renter-occupied housing units)
\item B25061 (universe: Vacant-for-rent and Rented, Not Occupied Housing Units)
}
}
