% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-calc-fncs.R
\name{get.cagr.comparison}
\alias{get.cagr.comparison}
\title{get.cagr.comparison}
\usage{
get.cagr.comparison(x)
}
\arguments{
\item{x}{dataframe with \code{geoid}, \code{variable}, \code{year}, and \code{estimate} columns.}
}
\description{
Relatively inflexible function that gets change, percent change, and CAGR
across regions and variables. Assumes columns \code{geoid}, \code{variable}, \code{year},
and \code{estimate}, as from data retrieved from tidycensus with minimal
processing (add year and rename tolower).
}
\details{
TODO make the columns variable.
}
