% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tract-attrs.R
\name{get.tract.attrs}
\alias{get.tract.attrs}
\title{get.tract.attrs}
\usage{
get.tract.attrs(
  state,
  cofps = NULL,
  year,
  geo = "tract",
  get.demographics.and.commute = F,
  keep.geos = F
)
}
\arguments{
\item{state}{state fp code}

\item{cofps}{county fp codes (3-character)}

\item{year}{year.}

\item{geo}{"tract" (default) or "block group" or "county", or other geo in
\code{censusrx::tigris.acs.geo.equivalents}}

\item{get.demographics.and.commute}{Whether to add some demographic and
commute information. These may be pulled separately with
\code{censusrx::tidycensus2recoded.tblList} -- b/c they don't have an obvious 1
row/tract, this would allow more specific analysis. When they are pulled
here they are pulled in wide, where the "transit" columns will refer to
combined commuters by active+public transit.}

\item{keep.geos}{Whether or not to keep geometries retrieved from \code{tigris}
function. They will be pulled initially regardless to get ALAND and
calculate densities.}
}
\description{
One massive function to pull a lot of frequently-used data points at most
common census geographies for a given year and set of counties. Will return a
one-row-by-areal-unit dataframe.
}
\details{
TODO: could be add an argument for keeping MOEs

Other acs pulls that are not 1-row by tract can be from
\code{censusrx::tidycensus2recoded.tblList}
}
