% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-mass-pulls.R
\name{gett.census.totals}
\alias{gett.census.totals}
\title{gett.census.totals}
\usage{
gett.census.totals(states, years, geo = "tract", cofps = NULL, survey = "acs5")
}
\arguments{
\item{states, geo, years}{passed to \code{tidycensus::get_acs}. There is one get_acs
call for every combination of table and years.}

\item{cofps}{countyfp codes passed to \code{tidycensus::get_acs.} Note may note
work with all geographies or when multiple states are queried. NULL by
get all geographies by type within \code{states}.}
}
\description{
Gets totals for area. Like population, households, housing units, and land
area. Wraps \code{tidycensus::get_acs}. 1 year at a time please.
}
