% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-mass-pulls.R
\name{mass.acs.pull.wrapper}
\alias{mass.acs.pull.wrapper}
\title{mass.acs.pull.tract.attrs}
\usage{
mass.acs.pull.wrapper(state, county, year)
}
\description{
Wrapps all the other ACS/tidycensus wrapper functions here to pull a
selection of tables with recoded values, medians, and totals, over a
selection of states and years.
}
\details{
You can use \code{taux::acs.tbl.index} for reminders of what each table is.
}
