% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-mass-pulls.R
\name{pull.tidycensus.median.tables}
\alias{pull.tidycensus.median.tables}
\title{pull.tidycensus.median.tables}
\usage{
\method{pull}{tidycensus.median.tables}(
  states = NULL,
  years,
  geo = "tract",
  tbls = c(med.hhinc = "B19013", med.crent = "B25058", med.hcosts = "B25105", med.hvalue
    = "B25077"),
  cofps = NULL,
  survey = "acs5"
)
}
\arguments{
\item{states, geo, years}{passed to \code{tidycensus::get_acs}. There is one get_acs
call for every combination of table and years.}

\item{cofps}{countyfp codes passed to \code{tidycensus::get_acs.} Note may note
work with all geographies or when multiple states are queried. NULL by
get all geographies by type within \code{states}.}
}
\description{
Pulls median tables, like hh income and contract rent:
}
\details{
\itemize{
\item B19013 (median household income; universe is households)
\item B25058 (median contract rent; universe is "Renter-occupied Housing Units Paying Cash Rent")
\item B25105 (median monthly housing costs; universe is "Occupied Housing Units With Monthly Housing Costs")
\item B25077 (median home value; universe is owner-occupied housing units)
}
}
