% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tract-attrs.R
\name{get.tract.attrs}
\alias{get.tract.attrs}
\title{get.tract.attrs}
\usage{
get.tract.attrs(state, cofps, year, geo = "tract")
}
\arguments{
\item{state}{state fp code}

\item{cofps}{county fp codes (3-character)}

\item{year}{year.}

\item{geo}{'tract' (default) or "block group"}
}
\description{
One massive function to pull a lot of frequently-used data points at the
tract or block group level for a given year and set of counties. Will return
a one-row-by tract (or BG) dataframe.
}
\details{
TODO: could be add an argument for keeping MOEs

Other acs pulls that are not 1-row by tract can be from
\code{censusrx::tidycensus2recoded.tblList}
}
