% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tract-attrs.R
\name{get.tract.attrs}
\alias{get.tract.attrs}
\title{get.tract.attrs}
\usage{
get.tract.attrs(
  state,
  cofps,
  year,
  geo = "tract",
  get.demographics.and.commute = F
)
}
\arguments{
\item{state}{state fp code}

\item{cofps}{county fp codes (3-character)}

\item{year}{year.}

\item{geo}{"tract" (default) or "block group"}

\item{get.demographics.and.commute}{Whether to add some demographic and
commute information. These may be pulled separately with
\code{censusrx::tidycensus2recoded.tblList} -- b/c they don't have an obvious 1
row/tract, this would allow more specific analysis. When they are pulled
here they are pulled in wide, where the "transit" columns will refer to
combined active+public transit.}
}
\description{
One massive function to pull a lot of frequently-used data points at the
tract or block group level for a given year and set of counties. Will return
a one-row-by tract (or BG) dataframe.
}
\details{
TODO: could be add an argument for keeping MOEs

Other acs pulls that are not 1-row by tract can be from
\code{censusrx::tidycensus2recoded.tblList}
}
