% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tract-attrs.R
\name{get.tract.attrs}
\alias{get.tract.attrs}
\title{get.tract.attrs}
\usage{
get.tract.attrs(state, cofps, yr)
}
\arguments{
\item{state}{state fp code}

\item{cofps}{county fp codes (3-character)}

\item{yr}{year.}
}
\description{
One massive function to pull a lot of frequently-used data points at the
tract level for a given year and set of counties. Will return a one-row-by
tract dataframe.
}
\details{
Other acs pulls that are not 1-row by tract can be from \code{censusrx::tidycensus2recoded.tblList}
}
