% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-mass-pulls.R
\name{pull.tidycensus.median.tables}
\alias{pull.tidycensus.median.tables}
\title{pull.tidycensus.median.tables}
\usage{
pull.tidycensus.median.tables(
  states,
  years,
  geo = "tract",
  tbls = c("B19013", "B25058", "B25077"),
  cofps = NULL,
  survey = "acs5"
)
}
\arguments{
\item{states, geo, years}{passed to \code{tidycensus::get_acs}. There is one get_acs
call for every combination of table and years.}

\item{cofps}{countyfp codes passed to \code{tidycensus::get_acs.} Note may note
work with all geographies or when multiple states are queried. NULL by
get all geographies by type within \code{states}.}
}
\description{
Pulls median tables, like hh income and contract rent:
\itemize{
\item B19013 (median household income; universe is households)
\item B25058 (median contract rent; universe is "Renter-occupied Housing Units Paying Cash Rent")
\item B25077 (median home value; universe is owner-occupied housing units)
}
}
