% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus-mass-pulls.R
\name{pull.tidycensus.median.tables}
\alias{pull.tidycensus.median.tables}
\title{pull.tidycensus.median.tables}
\usage{
\method{pull}{tidycensus.median.tables}(
  states,
  years,
  geo = "tract",
  tbls = c("B19013", "B25058", "B25077"),
  survey = "acs5"
)
}
\arguments{
\item{states}{vector of state FIPS or 2-letters}

\item{years}{vector of years}

\item{geo}{selection geography, passed onto \code{tidycensus}}
}
\description{
Pulls median tables, like hh income and contract rent:
\itemize{
\item B19013 (median household income; universe is households)
\item B25058 (median contract rent; universe is "Renter-occupied Housing Units Paying Cash Rent")
\item B25077 (median home value; universe is owner-occupied housing units)
}
}
