% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tigerline-metadata.R
\name{roads.tigerline.metadata}
\alias{roads.tigerline.metadata}
\title{roads.tigerline.metadata}
\usage{
roads.tigerline.metadata(rttyp = F)
}
\description{
Export metadata for roads tigerline features
}
